Discord Pipeline/requirements.txt
kafka-python
pyspark
psycopg2-binary
plotly
pandas
uvicorn
fastapi